Create pixel_canvas
Had been a great learning experience.
first of all, display_Set_mode() sets the window, so when we are calling the pixie, it is not needed to set the display. pixie is only a surface to be blit'ed on the canvas
then, pixie has its pixels numbers from 0,pixieside where as canvas has its pixels numbered from 0 to canvas side.
bliting is like stitching. once the pixie is blit'ed , the pixel can be accessed by referencing the canvas pixels.(need to check this further)